Git cleanup and added basic chrono-folding
A new parsing stage has been added, in the file
"src/parse/delay_modify.js"

The term chrono-folding is used, but basically it tracks constant changes to
registers, and instead of changing the register itself, changes the
instructions that these changes will affect.

This opens the door to faster execution, but more notably, a wider
potential for application of other optimizations.